+++ apps/router-e2e/__e2e__/modal/app/index.tsx
@@ -164,6 +164,7 @@ function FitForm() {
         visible={open}
         onClose={() => setIsOpen(false)}
         presentationStyle="formSheet"
+        cornerRadius={20}
         detents="fitToContents">
         <ScrollView
           style={{ gap: 8 }}